feat: delete 이후 UI 삭제 반영
